Time: 0 ms (100.00%) | Memory: 8.6 MB (30.23%) - LeetSync

+++ 4168-mirror-distance-of-an-integer/mirror-distance-of-an-integer.cpp
@@ -9,6 +9,6 @@ public:
             reverse *= copy ? 10 : 1;
         }
 
-        return abs((long)n - (long)(reverse));
+        return abs(n - reverse);
     }
 };
